feat: PackageModel Update

+++ src/models/PackageModel.py
@@ -94,7 +94,7 @@ class HeightAbsolute(Config):
 
 class MaskValue(Config):
     name: Literal["MaskValue"] = "MaskValue"
-    value: int = Field(ge=0.0, le=1.0 ,default=0.5)
+    value: float = Field(ge=0.0, le=1.0 ,default=0.5)
     type: Literal["number"] = "number"
     field: Literal["textInput"] = "textInput"
